move converter into the prometheus
