+++ pkg/tsdb/loki/api.go
@@ -20,10 +20,11 @@ import (
 	"go.opentelemetry.io/otel/trace"
 
 	"github.com/grafana/grafana-plugin-sdk-go/backend"
+
 	"github.com/grafana/grafana/pkg/infra/log"
 	"github.com/grafana/grafana/pkg/infra/tracing"
 	"github.com/grafana/grafana/pkg/tsdb/loki/instrumentation"
-	"github.com/grafana/grafana/pkg/util/converter"
+	"github.com/grafana/grafana/pkg/tsdb/prometheus/converter"
 )
 
 type LokiAPI struct {
@@ -334,7 +335,7 @@ func getSupportingQueryHeaderValue(req *http.Request, supportingQueryType Suppor
 		value = "datasample"
 	case SupportingQueryInfiniteScroll:
 		value = "infinitescroll"
-	default: //ignore
+	default: // ignore
 	}
 
 	return value
