+++ pkg/tsdb/prometheus/converter/testdata/loki-streams-a-frame.jsonc
