+++ pkg/tsdb/prometheus/converter/testdata/loki-streams-b-frame.jsonc
