+++ pkg/tsdb/prometheus/converter/testdata/loki-streams-c-frame.jsonc
