+++ pkg/tsdb/prometheus/converter/testdata/loki-streams-structured-metadata.json
