+++ pkg/tsdb/prometheus/converter/testdata/prom-error-frame.jsonc
