+++ pkg/tsdb/prometheus/converter/testdata/prom-exemplars-a-frame.json
