+++ pkg/tsdb/prometheus/converter/testdata/prom-exemplars-b.json
