+++ pkg/tsdb/prometheus/converter/testdata/prom-labels-frame.jsonc
