+++ pkg/tsdb/prometheus/converter/testdata/prom-matrix-histogram-no-labels.json
