+++ pkg/tsdb/prometheus/converter/testdata/prom-matrix-histogram-partitioned-frame.jsonc
