+++ pkg/tsdb/prometheus/converter/testdata/prom-scalar-frame.jsonc
