+++ pkg/tsdb/prometheus/converter/testdata/prom-scalar.json
