+++ pkg/tsdb/prometheus/converter/testdata/prom-string-frame.jsonc
