+++ pkg/tsdb/prometheus/converter/testdata/prom-string.json
