+++ pkg/tsdb/prometheus/converter/testdata/prom-vector-histogram-no-labels.json
